move tests inside a store folder, create state under selector
